+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # pgEdge Platform Release Notes #############
 
-## done for 24.1.5 as of 2024-01-02 ###########
+## done for 24.1.5 as of 2024-01-04 ###########
+  - remove support for spock31
+  - improve pg14 with spock32 patches
   - opened up for development